moved and adjusted login and data pages to Next.JS files
